Room added

+++ src/pages/AdminDashboard/components/room/AddRoomType.jsx
@@ -17,6 +17,17 @@ import { ToastAction } from "@radix-ui/react-toast";
 import axios from "axios";
 import { Trash2 } from "lucide-react";
 import { useEffect, useState } from "react";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 
 export function AddRoomType() {
   const { roomTypes, dispatch } = useRoomTypeContext();
@@ -39,41 +50,30 @@ export function AddRoomType() {
       });
 
       if (response.status === 201) {
-        setType('');
+        setType("");
         dispatch({ type: "CREATE_TYPE", payload: response.data });
-        setResponseG(response.data);
+        setResponseG(response.data.message);
       }
     } catch (err) {
-      if (err.response) {
-        if (err.response.status === 409) {
-          setError(err.response.data.error);
-        } else {
-          setError("An error occurred. Please try again later.");
-        }
-      } else if (err.request) {
-        setError(
-          "No response received. Please check your internet connection."
-        );
-      } else {
-        setError("An error occurred. Please try again later.");
-      }
+      setError(err.response.data.error)
       setIsLoading(false);
     }
   };
 
-  const DeleteRoomType = async (id)=>{
+  const DeleteRoomType = async (id) => {
+    setResponseG(null)
+    setError(null);
+    try {
+      const response = await axios.delete("/api/roomType/delete/" + id);
 
-    try{
-const response = await axios.delete("/api/roomType/delete/"+id)
-
-if(response.status === 200){
-  dispatch({type: "DELETE_TYPE", payload: response.data})
-}
-    }catch(err){
-console.log(err)
+      if (response.status === 200) {
+        dispatch({ type: "DELETE_TYPE", payload: response.data });
+        setResponseG(response.data.message)
+      }
+    } catch (err) {
+      setError(err.response.data.error)
     }
-
-  }
+  };
 
   useEffect(() => {
     if (error) {
@@ -81,12 +81,6 @@ console.log(err)
         variant: "destructive",
         title: "Uh oh! Something went wrong.",
         description: `${error}`,
-        action: (
-          <ToastAction altText="Try again" onClick={AddType}>
-            Try again
-          </ToastAction>
-        ),
-        position: "center",
       });
     }
   }, [error]);
@@ -94,7 +88,7 @@ console.log(err)
   useEffect(() => {
     if (responseG) {
       toast({
-        title: `${responseG.roomType.type} Room Type added successfully!`,
+        title: `${responseG}!`,
       });
     }
   }, [responseG]);
@@ -110,7 +104,7 @@ console.log(err)
             <div className="grid gap-2">
               <Label>Room Type</Label>
               <Input
-              value={type}
+                value={type}
                 onChange={(e) => setType(e.target.value)}
                 type="text"
                 placeholder="Suite"
@@ -136,13 +130,34 @@ console.log(err)
                 >
                   {type.type}
 
-                  <Button
-                    variant="outline"
-                  onClick={()=>DeleteRoomType(type._id)}
-                    size="icon"
-                  >
-                    <Trash2 className="h-4 w-4" />
-                  </Button>
+                  <AlertDialog>
+                    <AlertDialogTrigger asChild>
+                      <Button variant="outline" size="icon">
+                        {" "}
+                        <Trash2 className="h-4 w-4" />
+                      </Button>
+                    </AlertDialogTrigger>
+                    <AlertDialogContent>
+                      <AlertDialogHeader>
+                        <AlertDialogTitle>
+                          Are you absolutely sure?
+                        </AlertDialogTitle>
+                        <AlertDialogDescription>
+                          This action cannot be undone. Deleting this room type
+                          will remove all associated rooms. Are you sure you
+                          want to proceed?
+                        </AlertDialogDescription>
+                      </AlertDialogHeader>
+                      <AlertDialogFooter>
+                        <AlertDialogCancel>Cancel</AlertDialogCancel>
+                        <AlertDialogAction
+                          onClick={() => DeleteRoomType(type._id)}
+                        >
+                          Continue
+                        </AlertDialogAction>
+                      </AlertDialogFooter>
+                    </AlertDialogContent>
+                  </AlertDialog>
                 </div>
                 <Separator className="my-2" />
               </>
